fixing problems from [jit.pix]'s incompatibility with 1-plane matrices
Also, since it doesn't really make sense to have the centroid snake drawing in greyscale, from now on the history video is always 4-plane (ie. the snake dots are always red), and because of this, every matrix is converted to 4-plane for the real-time preview (won't be visible, but [jit.glue] won't complain.